+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Unreleased
     - (BREAKING CHANGE) Fixed bug in CapiModel where change events were not being fired for keys added after model instantiation
     listeners registered with 'on' in CapiModel
-    now receive parameters [model, <value of key that changed>] (how Backbone does it)
+    now receive parameters [model, < value of key that changed >] (how Backbone does it)
     used to receive [model, model.attributes]
 
 - 2.0.0 (10 Jan 2017)