+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Release History
 
-- Unreleased
+- 3.0.1 (25 Jan 2017)
     - Fixed SimCapiValue parsing empty array as an array with one empty string element
 - 3.0.0 (16 Jan 2017)
     - Fixed bug in CapiModel where change events were not being